+++ packages/web3-utils/src/string_manipulation.ts
