+++ packages/web3-utils/test/fixtures/string_manipulation.ts
